# Conflicts:
#	src/main/kotlin/com/prasetia/erp/controller/web/CmeController.kt
#	src/main/kotlin/com/prasetia/erp/pojo/cme/CmeSummaryYearProjectTypeCustData.kt

+++ src/main/kotlin/com/prasetia/erp/pojo/PreventiveCustomer.kt
@@ -13,27 +13,30 @@ class PreventiveCustomerGroup(
         val id:Long,
         var customer: String?,
         var customer_id: Long?,
-        val nilai_po: Long?,
-        val nilai_penagihan: Long?,
-        val nilai_budget: Long?,
-        val realisasi_budget: Long?,
+        val nilai_po: Double?,
+        val nilai_penagihan: Double?,
+        val nilai_budget: Double?,
+        val realisasi_budget: Double?,
         var detail: MutableList<PreventiveCustomerGroupDetail>?
 ){
-    constructor(): this(0, "", 0, 0,0,0,0, null)
+    constructor(): this(0, "", 0, 0.0,0.0,
+            0.0,0.0, null)
 }
 
 class PreventiveCustomerGroupDetail(
         val id: Long,
         val area: String?,
-        val nilai_po: Long?,
-        val nilai_penagihan: Long?,
-        val nilai_budget: Long?,
-        val realisasi_budget: Long?,
-        val laba_rugi: Long?,
-        val persent_penagihan:Long?,
-        val persent_budget:Long?,
-        val persent_laba_rugi:Long?,
+        val nilai_po: Double?,
+        val nilai_penagihan: Double?,
+        val nilai_budget: Double?,
+        val realisasi_budget: Double?,
+        val laba_rugi: Double?,
+        val persent_penagihan:Double?,
+        val persent_budget:Double?,
+        val persent_laba_rugi:Double?,
         val area_id:Long?
 ){
-    constructor(): this(0, "", 0, 0, 0, 0, 0, 0, 0, 0, 0)
+    constructor(): this(0, "", 0.0, 0.0,
+            0.0, 0.0, 0.0,
+            0.0, 0.0, 0.0, 0)
 }
