+++ src/main/kotlin/com/prasetia/erp/pojo/cme/CmeSummaryYearProjectTypeCustData.kt
@@ -5,9 +5,9 @@ class CmeInvoiceProjectData(
         val project_id:Long?,
         val name:String?,
         val state:String?,
-        val nilai_invoice:Long?
+        val nilai_invoice:Double?
 ){
-    constructor(): this(0,0,"","",0)
+    constructor(): this(0,0,"","",0.0)
 }
 
 class CmeSummaryYearProjectTypeCustData(
@@ -16,25 +16,27 @@ class CmeSummaryYearProjectTypeCustData(
         val jumlah_site: Long,
         val project_type: String,
         val site_cancel: Long,
-        val nilai_po: Long,
-        val nilai_invoice: Long,
-        val nilai_budget: Long,
-        val realisasi_budget: Long,
-        val estimate_po: Long,
+        val nilai_po: Double,
+        val nilai_invoice: Double,
+        val nilai_budget: Double,
+        val realisasi_budget: Double,
+        val estimate_po: Double,
         val site_type_id: Long?,
         val customer:String?,
         val customer_id:Long?,
-        val percentage: Float,
-        val remaining_invoice: Long,
-        val percentage_realization: Float,
-        val profit_loss: Long,
-        val percentage_profit_realization: Float,
-        val percentage_profit_po: Float,
+        val percentage: Double,
+        val remaining_invoice: Double,
+        val percentage_realization: Double,
+        val profit_loss: Double,
+        val percentage_profit_realization: Double,
+        val percentage_profit_po: Double,
         var project_list: List<CmeYearProjectTypeCustProjectDetailData>?
 ){
-    constructor(): this(0,0,0,"",0,0,0,0,
-            0,0, 0, "", 0, 0f, 0,
-            0f, 0, 0f, 0f, null)
+    constructor(): this(0,0,0,"",0,
+            0.0,0.0,0.0,
+            0.0,0.0, 0, "",
+            0, 0.0, 0.0,
+            0.0, 0.0, 0.0, 0.0, null)
 }
 class CmeYearProjectTypeCustProjectDetailData(
     val id: Long,
@@ -42,20 +44,20 @@ class CmeYearProjectTypeCustProjectDetailData(
     val year_project: Long,
     val project_type: String,
     val project_id: String,
-    val nilai_po: Long,
+    val nilai_po: Double,
     val no_po: String?,
-    val nilai_invoice: Long,
-    val nilai_budget: Long,
-    val realisasi_budget: Long,
-    val estimate_po: Long,
+    val nilai_invoice: Double,
+    val nilai_budget: Double,
+    val realisasi_budget: Double,
+    val estimate_po: Double,
     val customer:String?,
     val customer_id:Long?,
     val site_type_id: Long?,
-    var percent_po : Float?,
-    var percent_labarugi : Float?,
+    var percent_po : Double?,
+    var percent_labarugi : Double?,
     val area: String?,
     var invoice_list:List<CmeInvoiceProjectData>?
 ){
-    constructor(): this(0,"",0,"","",0,"",
-            0,0,0,0,"", 0, 0, 0f, 0f, "",null)
+    constructor(): this(0,"",0,"","",0.0,"",
+            0.0,0.0,0.0,0.0,"", 0, 0, 0.0, 0.0, "", null)
 }
