+++ src/main/kotlin/com/prasetia/erp/pojo/department/DepartmentBudgetData.kt
@@ -9,9 +9,9 @@ class DepartmentBudgetRealisasiData(
         val parent_id:Long,
         val ref:String?,
         val narration:String?,
-        val budget_realisasi:Long?
+        val budget_realisasi:Double?
 ){
-    constructor(): this(0,null,0,0, "","",0)
+    constructor(): this(0,null,0,0, "","",0.0)
 }
 
 class DepartmentBudgetDetailData(
@@ -20,33 +20,33 @@ class DepartmentBudgetDetailData(
         val line_id:Long?,
         val code:String?,
         val budget_item_view:String?,
-        val nilai_budget:Long?,
-        val realisasi_budget:Long?,
-        val persent_budget: Float?,
+        val nilai_budget:Double?,
+        val realisasi_budget:Double?,
+        val persent_budget: Double?,
         var realisasi:List<DepartmentBudgetRealisasiData>?
 ){
-    constructor(): this(0,0,0,"","",0,0,0f,null)
+    constructor(): this(0,0,0,"","",0.0,0.0,0.0,null)
 }
 
 class DepartmentBudgetYearData(
         val id:Long?,
         val department_name:String?,
         val department_id:Long?,
-        val nilai_budget:Long?,
-        val realisasi_budget:Long?,
-        val persent_budget: Float?,
+        val nilai_budget:Double?,
+        val realisasi_budget:Double?,
+        val persent_budget: Double?,
         var department_budget: MutableList<DepartmentBudgetData>?
 ){
-    constructor(): this(0,"", 0,0,0,0f, null)
+    constructor(): this(0,"", 0,0.0,0.0,0.0, null)
 }
 
 class DepartmentBudgetData(
         val id:Long?,
         val name:String?,
-        val nilai_budget:Long?,
-        val realisasi_budget: Long?,
-        val persent_budget: Float?,
+        val nilai_budget:Double?,
+        val realisasi_budget: Double?,
+        val persent_budget: Double?,
         var budget_detail: MutableList<DepartmentBudgetDetailData>?
 ){
-    constructor(): this(0,"",0,0,0f,null)
+    constructor(): this(0,"",0.0,0.0,0.0,null)
 }
