+++ src/main/kotlin/com/prasetia/erp/pojo/department/DepartmentYearData.kt
@@ -5,9 +5,9 @@ class DepartmentYearData(
         val id:Long,
         val department_name:String?,
         val department_id:Long?,
-        val nilai_budget:Long?,
-        val realisasi_budget:Long?,
-        val persent_budget: Float?
+        val nilai_budget:Double?,
+        val realisasi_budget:Double?,
+        val persent_budget: Double?
 ){
-    constructor(): this(0,"",0,0,0,0f)
+    constructor(): this(0,"",0,0.0,0.0,0.0)
 }
